+++ src/pages/ProfessionalHistory.test.tsx
@@ -1,17 +1,17 @@
 import { render, screen } from "@testing-library/react";
-import { Projects } from "./Projects";
+import { ProfessionalHistory } from "./ProfessionalHistory";
 
-describe("Projects", () => {
+describe("ProfessionalHistory", () => {
   it("renders h2", () => {
-    render(<Projects />);
+    render(<ProfessionalHistory />);
     const headings = screen.getAllByRole("heading");
     const h2 = headings[0];
     expect(h2).toBeInTheDocument();
-    expect(h2).toHaveTextContent("Projects");
+    expect(h2).toHaveTextContent("Professional History");
   });
 
   it("renders project data", () => {
-    render(<Projects />);
+    render(<ProfessionalHistory />);
 
     // Verifies multiple Projects are returned, but these are intentionally hard-coded here and may not align with the entire response over time
     expect(screen.getByText("Contract Consultant")).toBeDefined();
@@ -19,6 +19,5 @@ describe("Projects", () => {
       screen.getByText("Founder & Leader of Front-End Guild")
     ).toBeDefined();
     expect(screen.getByText("IoT Tech Lead")).toBeDefined();
-    expect(screen.getByText("Learnalotl")).toBeDefined();
   });
 });
