+++ src/pages/Projects.tsx
@@ -1,30 +1,28 @@
+import Carousel from "../components/Carousel";
 import { ProjectItem } from "../components/ProjectItem";
+import ThreeDCarousel from "../components/ThreeDCarousel";
+
 import { useProjects } from "../hooks/useProjects";
-import useFadeIn from "../hooks/useFadeIn";
 
 export const Projects = () => {
-  useFadeIn();
-
   const projectData = useProjects();
+
+  return <ThreeDCarousel />;
 
   return (
     <div className="flex flex-col m-auto gap-6 min-h-screen p-2 md:px-6 md:py-2 bg-brand-gradient-2">
       <div className="container-card">
         <h2 className="page-heading">Projects</h2>
 
-        <div className="flex flex-col gap-8">
-          {projectData.map((project, index) => {
-            const data = {
-              ...project,
-              imagePositionStart: Boolean(index % 2),
-            };
+        <Carousel>
+          {projectData.map((project) => {
             return (
               <div className="frosted-glass" key={project.name}>
-                <ProjectItem key={project.name} {...data} />
+                <ProjectItem {...project} />
               </div>
             );
           })}
-        </div>
+        </Carousel>
       </div>
     </div>
   );
